+++ src/mlpro/rl/examples/howto_rl_agent_001_run_agent_with_own_policy_on_gym_environment.py
@@ -21,10 +21,11 @@
 ## -- 2022-11-02  1.2.6     DA       Refactoring 
 ## -- 2022-11-07  1.3.0     DA       Refactoring 
 ## -- 2023-01-14  1.3.1     MRD      Removing default parameter new_step_api and render_mode for gym
+## -- 2023-04-19  1.3.2     MRD      Refactor module import gym to gymnasium
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.3.1 (2023-01-14)
+Ver. 1.3.2 (2023-04-19)
 
 This module shows how to run an own policy inside the standard agent model with an OpenAI Gym environment using 
 MLPro framework.
@@ -46,8 +47,8 @@ You will learn:
 
 from mlpro.bf.math import *
 from mlpro.rl import *
-from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
-import gym
+from mlpro.wrappers.gymnasium import WrEnvGYM2MLPro
+import gymnasium as gym
 import random
