+++ src/mlpro/rl/examples/howto_rl_agent_002_train_agent_with_own_policy_on_gym_environment.py
@@ -24,10 +24,11 @@
 ## -- 2022-11-02  1.3.6     DA       Refactoring 
 ## -- 2022-11-07  1.4.0     DA       Refactoring 
 ## -- 2023-01-14  1.4.1     MRD      Removing default parameter new_step_api and render_mode for gym
+## -- 2023-04-19  1.4.2     MRD      Refactor module import gym to gymnasium
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.4.1 (2023-01-14)
+Ver. 1.4.2 (2023-04-19)
 
 This module shows how to train an agent with a custom policy inside on an OpenAI Gym environment using
 MLPro framework.
@@ -49,8 +50,8 @@ You will learn:
 
 from mlpro.bf.math import *
 from mlpro.rl import *
-from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
-import gym
+from mlpro.wrappers.gymnasium import WrEnvGYM2MLPro
+import gymnasium as gym
 import random
 from pathlib import Path
