+++ src/mlpro/rl/examples/howto_rl_agent_011_train_and_reload_single_agent_gym.py
@@ -22,10 +22,11 @@
 ## -- 2023-03-02  1.2.1     LSB      Refactoring
 ## -- 2023-03-04  1.3.0     DA       Renamed
 ## -- 2023-03-27  1.3.1     DA       Refactoring
+## -- 2023-04-19  1.3.2     MRD      Refactor module import gym to gymnasium
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.3.1 (2023-03-27)
+Ver. 1.3.2 (2023-04-19)
 
 This module shows how to train a single agent and load it again to do some extra cycles.
 
@@ -40,10 +41,10 @@ You will learn:
 """
 
 
-import gym
+import gymnasium as gym
 from stable_baselines3 import PPO
 from mlpro.rl import *
-from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
+from mlpro.wrappers.gymnasium import WrEnvGYM2MLPro
 from mlpro.wrappers.sb3 import WrPolicySB32MLPro
 from pathlib import Path
