+++ src/mlpro/rl/examples/howto_rl_att_001_train_and_reload_single_agent_gym_sd.py
@@ -7,10 +7,11 @@
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
 ## -- 2023-03-04  1.0.0     DA       Creation as derivate of howto_rl_agent_011
 ## -- 2023-03-27  1.0.1     DA       Refactoring
+## -- 2023-04-19  1.0.2     MRD      Refactor module import gym to gymnasium
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.0.1 (2023-03-27)
+Ver. 1.0.2 (2023-04-19)
 
 As in Howto RL AGENT 011, this module shows how to train a single agent and load it again to do some
 extra cycles. In opposite to howto 011, stagnation detection is used to automatically end the
@@ -29,10 +30,10 @@ You will learn:
 """
 
 
-import gym
+import gymnasium as gym
 from stable_baselines3 import PPO
 from mlpro.rl import *
-from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
+from mlpro.wrappers.gymnasium import WrEnvGYM2MLPro
 from mlpro.wrappers.sb3 import WrPolicySB32MLPro
 from pathlib import Path
