+++ test/test_buffer.py
@@ -8,10 +8,11 @@
 ## -- 2021-09-27  1.0.0     WB       Creation
 ## -- 2021-09-27  1.0.0     WB       Release First Version
 ## -- 2022-11-07  1.1.0     DA       Refactoring
+## -- 2023-04-19  1.1.1     MRD      Refactor module import gym to gymnasium
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.1.0 (2022-11-07)
+Ver. 1.1.1 (2023-04-19)
 
 Unit test classes for SARBuffer.
 """
@@ -24,11 +25,11 @@ from mlpro.bf.various import *
 from mlpro.bf.math import *
 from mlpro.bf.ml import *
 from mlpro.rl.models import *
-from mlpro.wrappers.openai_gym import WrEnvGYM2MLPro
+from mlpro.wrappers.gymnasium import WrEnvGYM2MLPro
 from mlpro.rl.pool.sarsbuffer.PrioritizedBuffer import PrioritizedBuffer
 from mlpro.rl.pool.sarsbuffer.RandomSARSBuffer import RandomSARSBuffer
 from mlpro.rl.pool.policies.dummy import MyDummyPolicy
-import gym
+import gymnasium as gym
 import random
 from pathlib import Path
